+++ tests/it/io/json/mod.rs
@@ -1,0 +1,1 @@
+mod read;
