+++ tests/it/io/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(feature = "io_print")]
 mod print;
+
+#[cfg(feature = "io_json")]
+mod json;
